arq: fix [RSP-TIMEOUT] diagnostic field alignment (double passed to %d)
The RESPONDER receive-window diagnostic formatted time_left_to_send_last_frame
(a double) with %d. On the x86-64 SysV ABI a floating-point vararg is passed in
an XMM register, not an integer register, so the six integer %d conversions
consumed one integer-register argument too few: every field after it printed one
slot to the left and the final timeout= read an unpushed stack slot (stale
garbage, sometimes negative). The value armed into receiving_timeout was always
well-formed -- only the diagnostic was wrong, and it mislabelled the armed value
under sack= while timeout= showed garbage.

Cast the argument to (int) so the format and the argument agree; all six fields
now line up and timeout= reports the value actually armed. Print-only change: no
consumer read the mis-formatted value, so no runtime behaviour changes. Guarded
by the test_rsp_timeout_format regression added alongside.

+++ source/datalink_layer/arq_common.cc
@@ -2681,7 +2681,7 @@ void cl_arq_controller::calculate_receiving_timeout()
 		int rsp_timeout = (data_batch_size)*message_transmission_time_ms+time_left_to_send_last_frame+ptt_on_delay_ms;
 		// RSP timeout: base timeout covers batch reception
 		printf("[RSP-TIMEOUT] batch=%d msg_time=%d time_left=%d ptt=%d sack=%d -> timeout=%d\n",
-			data_batch_size, message_transmission_time_ms, time_left_to_send_last_frame, ptt_on_delay_ms, sack_enabled ? 1 : 0, rsp_timeout);
+			data_batch_size, message_transmission_time_ms, (int)time_left_to_send_last_frame, ptt_on_delay_ms, sack_enabled ? 1 : 0, rsp_timeout);
 		fflush(stdout);
 		set_receiving_timeout(rsp_timeout);
 	}
